restructure library folder structer per Rowdy's spec
