+++ libs/ui/src/components/atoms/index.ts
@@ -1,0 +1,1 @@
+export * from './input';
