+++ libs/ui/src/components/atoms/input/index.ts
@@ -1,0 +1,2 @@
+export * from './input';
+export * from './input.stories';
