--- libs/ui/src/global/index.ts
@@ -1,1 +1,0 @@
-export * from './globalStyles';
