+++ libs/ui/src/index.ts
@@ -1,2 +1,1 @@
-export * from './global';
 export * from './theme';
