--- libs/ui/src/theme/global/color.ts
@@ -1,6 +1,0 @@
-import { skyDark } from '@radix-ui/colors';
-
-export const Color = {
-  BgColor: skyDark.sky1,
-  FontColor: skyDark.sky12,
-};
