+++ libs/ui/src/theme/global/index.ts
@@ -1,2 +1,1 @@
-export * from './color';
 export * from './font';
